refactor(shared): move model types into types/base
