+++ server/src/models/users.model.ts
@@ -2,7 +2,7 @@ import { Sequelize, DataTypes, Model, ModelCtor } from 'sequelize'
 
 import { Tag } from './tags.model'
 import { IMinimatch } from 'minimatch'
-import { User as UserBaseDto } from '../../../shared/types'
+import { User as UserBaseDto } from '../../../shared/types/base'
 
 const USER_MODEL_NAME = 'user'
